# backend/services/saved-service/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Inflap Saved Service API
  version: 1.0.0
  description: |
    Owner-scoped HTTP contract for Saved items and private collections.
    The authenticated owner is derived only from the verified session; request
    paths, query parameters, and bodies never accept an owner identifier.
    All personal responses are private and non-cacheable.
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/users/me/saved-items/query:
    parameters:
      - $ref: '#/components/parameters/ClientPlatformHeader'
      - $ref: '#/components/parameters/AppBuildHeader'
      - $ref: '#/components/parameters/AcceptLanguage'
    post:
      tags: [Saved items]
      summary: Search saved items
      description: |
        Searches only the authenticated owner's Saved projection. Cursor scope
        includes the normalized effective locale and all filters. Raw search
        text and cursor plaintext must not be logged.
      operationId: querySavedItems
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedItemsQueryRequest'
      responses:
        '200':
          $ref: '#/components/responses/SavedItemsSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Verified user access token with a current login-session generation.
  parameters:
    ClientPlatformHeader:
      name: X-Client-Platform
      in: header
      required: false
      description: |
        Bounded mobile platform input for owner-sticky product rollout. It is
        not an authentication or authorization claim. Missing or invalid input
        disables new expansion but never blocks canonical reads or reduction.
      schema:
        type: string
        enum: [android, ios]
    AppBuildHeader:
      name: X-App-Build
      in: header
      required: false
      description: |
        Positive decimal store build used only by product rollout minimum-build
        rules. It grants no data access. Missing or invalid input disables new
        expansion but never blocks canonical reads or reduction.
      schema:
        type: integer
        format: uint64
        minimum: 1
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: |
        Requested Saved presentation locale. The service normalizes the value to
        one effective EN, RU, or KK locale and returns it in Content-Language.
      schema:
        $ref: '#/components/schemas/AcceptLanguageValue'
...
  headers:
    PrivateCacheControl:
      description: Personal responses must never be stored by shared or private caches.
      required: true
      schema:
        type: string
        const: 'private, no-store'
    PragmaNoCache:
      description: Compatibility directive for legacy intermediaries.
      required: true
      schema:
        type: string
        const: no-cache
    ContentLanguage:
      description: Effective normalized Saved presentation locale.
      required: true
      schema:
        $ref: '#/components/schemas/Locale'
    VaryPersonal:
      description: Inputs that can change an authenticated localized response.
      required: true
      schema:
        type: string
        const: 'Authorization, Accept-Language'
    RequestId:
      description: Sanitized request correlation identifier.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    RetryAfter:
      description: Minimum whole seconds before polling or retrying.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 30
    WWWAuthenticate:
      description: Bearer authentication challenge.
      required: true
      schema:
        type: string
        pattern: '^Bearer(?: .*)?$'
  responses:
...
    SavedItemsSearchResponse:
      description: Owner-scoped Saved search page without aggregate counters.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SavedItemsSearchPage'
...
    BadRequestError:
      description: Request syntax, encoding, cursor, filter, or validation failure.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnauthorizedError:
      description: The verified authenticated session is absent or invalid.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ForbiddenError:
      description: Personal-data access is blocked by the shared platform policy.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFoundError:
      description: Neutral not-found response that does not reveal foreign personal resources.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    TooManyRequestsError:
      description: Subject, IP, operation, or owner-quota rate control rejected the request.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ServiceUnavailableError:
      description: Saved or a required source dependency is temporarily unavailable.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    Locale:
      type: string
      description: Effective Saved locale (EN, RU, or KK).
      enum: [en, ru, kk]
    AcceptLanguageValue:
      type: string
      minLength: 1
      maxLength: 128
      description: Standard Accept-Language value; the service selects one effective en, ru, or kk locale.
    EntityType:
      type: string
      enum: [ATTRACTION, ACTIVITY, USER, POST]
    EntityKey:
      type: string
      minLength: 1
      maxLength: 512
      x-max-utf8-bytes: 512
      description: Opaque canonical source identifier; semantics are source-owned.
    CollectionId:
      type: string
      format: uuid
...
    Version:
      type: integer
      format: int64
      minimum: 0
      maximum: 9223372036854775807
      description: Server-issued monotonic version, comparable only within its typed resource scope.
    PositiveVersion:
      allOf:
        - $ref: '#/components/schemas/Version'
        - minimum: 1
    RelationshipGeneration:
      type: string
      format: uuid
      description: Random server-issued identity of one relationship activation generation.
...
    SearchText:
      type: string
      minLength: 1
      maxLength: 200
      description: Saved search text, bounded to 200 Unicode code points and a server token limit.
    OpaqueCursor:
      type: string
      minLength: 1
      maxLength: 4096
      description: |
        Opaque AEAD-protected continuation token bound to owner, scope, filters,
        effective locale, key ID, and expiry. Clients must not parse or log it.
    SavedTarget:
      type: object
      additionalProperties: false
      required: [entity_type, entity_id]
      properties:
        entity_type:
          $ref: '#/components/schemas/EntityType'
        entity_id:
          $ref: '#/components/schemas/EntityKey'
    SavedItemsQueryRequest:
      type: object
      additionalProperties: false
      required: [search]
      properties:
        type:
          $ref: '#/components/schemas/EntityType'
        collection_id:
          $ref: '#/components/schemas/CollectionId'
        uncollected:
          type: boolean
          default: false
        search:
          $ref: '#/components/schemas/SearchText'
        cursor:
          $ref: '#/components/schemas/OpaqueCursor'
        limit:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
          default: 30
      allOf:
        - not:
            required: [collection_id, uncollected]
            properties:
              uncollected:
                const: true
...
    AvailableSavedCardProjection:
      type: object
      additionalProperties: false
      required:
        - projection_version
        - content_state
        - display_locale
        - title
        - canonical_detail_route
      properties:
        projection_version:
          $ref: '#/components/schemas/Version'
        content_state:
          const: AVAILABLE
        display_locale:
          $ref: '#/components/schemas/Locale'
        title:
          type: string
          minLength: 1
          maxLength: 300
        subtitle:
          type: string
          maxLength: 500
        image_url:
          type: string
          format: uri
          maxLength: 2048
        canonical_detail_route:
          type: string
          minLength: 1
          maxLength: 256
          pattern: '^/(places|activities)/[^/?#%\\]+$|^/users/[^/?#%\\]+/profile$'
          description: |
            Source-validated first-party GoRouter location for this exact
            immutable target. It is present only for AVAILABLE PUBLIC content;
            clients must not derive a route from entity type or target ID.
        source_updated_at:
          type: string
          format: date-time
    UnavailableSavedCardProjection:
      type: object
      additionalProperties: false
      required: [projection_version, content_state]
      properties:
        projection_version:
          $ref: '#/components/schemas/Version'
        content_state:
          const: UNAVAILABLE
    SavedCardProjection:
      description: PUBLIC card payload or a payload-free unavailable placeholder.
      oneOf:
        - $ref: '#/components/schemas/AvailableSavedCardProjection'
        - $ref: '#/components/schemas/UnavailableSavedCardProjection'
      discriminator:
        propertyName: content_state
        mapping:
          AVAILABLE: '#/components/schemas/AvailableSavedCardProjection'
          UNAVAILABLE: '#/components/schemas/UnavailableSavedCardProjection'
    ActiveRelationship:
      type: object
      additionalProperties: false
      required: [generation, version, saved_at]
      properties:
        generation:
          $ref: '#/components/schemas/RelationshipGeneration'
        version:
          $ref: '#/components/schemas/PositiveVersion'
        saved_at:
          type: string
          format: date-time
        attribution_id:
          type: string
          format: uuid
          description: Present only when relationship-level analytics consent allows it.
...
    SearchMatch:
      type: object
      additionalProperties: false
      required: [match_rank, match_kind, matched_field, matched_locale]
      properties:
        match_rank:
          type: integer
          format: int32
          minimum: 0
          maximum: 1000000
        match_kind:
          type: string
          enum: [EXACT, TOKEN, PREFIX]
        matched_field:
          type: string
          enum: [TITLE, CITY, COUNTRY]
        matched_locale:
          $ref: '#/components/schemas/Locale'
        alternate_public_display_value:
          type: string
          maxLength: 300
          description: Present only when the match differs from the current display field.
    SavedSearchItem:
      type: object
      additionalProperties: false
      required: [target, relationship, effective_collection_count, projection, match]
      properties:
        target:
          $ref: '#/components/schemas/SavedTarget'
        relationship:
          $ref: '#/components/schemas/ActiveRelationship'
        effective_collection_count:
          type: integer
          format: int32
          minimum: 0
          maximum: 200
        projection:
          $ref: '#/components/schemas/SavedCardProjection'
        match:
          $ref: '#/components/schemas/SearchMatch'
    QuotaWarning:
      type: object
      additionalProperties: false
      required: [resource, limit, remaining]
      properties:
        resource:
          type: string
          enum:
            - SAVED_ITEMS
            - SAVED_COLLECTIONS
            - COLLECTION_MEMBERSHIPS
            - OWNER_MEMBERSHIPS
        limit:
          type: integer
          format: int64
          minimum: 1
        remaining:
          type: integer
          format: int64
          minimum: 0
      description: Returned only near an authoritative quota threshold or after rejection.
...
    SavedItemsSearchPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor, has_more]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/SavedSearchItem'
        next_cursor:
          oneOf:
            - $ref: '#/components/schemas/OpaqueCursor'
            - type: 'null'
        has_more:
          type: boolean
        quota_warning:
          $ref: '#/components/schemas/QuotaWarning'
...
    ErrorCode:
      type: string
      enum:
        - INVALID_ARGUMENT
        - UNAUTHENTICATED
        - NOT_FOUND
        - SAVED_TARGET_TYPE_UNSUPPORTED
        - SAVED_TARGET_UNAVAILABLE
        - SAVED_DEPENDENCY_UNAVAILABLE
        - SAVED_MUTATION_STALE
        - SAVED_MUTATION_REPLAY_MISMATCH
        - SAVED_COLLECTION_NOT_FOUND
        - SAVED_COLLECTION_DELETED
        - SAVED_COLLECTION_TITLE_INVALID
        - SAVED_COLLECTION_TITLE_CONFLICT
        - SAVED_COLLECTION_LIMIT_REACHED
        - SAVED_COLLECTION_ITEM_LIMIT_REACHED
        - SAVED_MEMBERSHIP_LIMIT_REACHED
        - SAVED_ITEM_LIMIT_REACHED
        - SAVED_REQUEST_IN_PROGRESS
        - SAVED_OPERATION_EXPIRED
        - SAVED_CURSOR_INVALID
        - SAVED_RATE_LIMITED
        - SAVED_TEMPORARILY_UNAVAILABLE
        - PLATFORM_PERSONAL_DATA_LOCKED
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [code, retryable]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        retryable:
          type: boolean
        retry_after_ms:
          type: integer
          format: int64
          minimum: 0
          maximum: 86400000
        request_id:
          type: string
          minLength: 1
          maxLength: 128